Add maxGlobalLongSize check
