+++ src/Api/index.js
@@ -186,50 +186,6 @@ export function useAllOrdersStats(chainId) {
   return res ? res.data.orderStat : null;
 }
 
-export function useInfoTokens(library, chainId, active, tokenBalances, fundingRateInfo, vaultPropsLength) {
-  const tokens = getTokens(chainId);
-  const vaultReaderAddress = getContract(chainId, "VaultReader");
-  const vaultAddress = getContract(chainId, "Vault");
-  const positionRouterAddress = getContract(chainId, "PositionRouter");
-  const nativeTokenAddress = getContract(chainId, "NATIVE_TOKEN");
-
-  const whitelistedTokens = getWhitelistedTokens(chainId);
-  const whitelistedTokenAddresses = whitelistedTokens.map((token) => token.address);
-
-  const { data: vaultTokenInfo } = useSWR(
-    [`useInfoTokens:${active}`, chainId, vaultReaderAddress, "getVaultTokenInfoV3"],
-    {
-      fetcher: fetcher(library, VaultReader, [
-        vaultAddress,
-        positionRouterAddress,
-        nativeTokenAddress,
-        expandDecimals(1, 18),
-        whitelistedTokenAddresses,
-      ]),
-    }
-  );
-
-  const indexPricesUrl = getTracerServerUrl(chainId, "/prices");
-  const { data: indexPrices } = useSWR([indexPricesUrl], {
-    fetcher: (...args) => fetch(...args).then((res) => res.json()),
-    refreshInterval: 500,
-    refreshWhenHidden: true,
-  });
-
-  return {
-    infoTokens: getInfoTokens(
-      tokens,
-      tokenBalances,
-      whitelistedTokens,
-      vaultTokenInfo,
-      fundingRateInfo,
-      vaultPropsLength,
-      indexPrices,
-      nativeTokenAddress
-    ),
-  };
-}
-
 export function useUserStat(chainId) {
   const query = gql(`{
     userStat(id: "total") {
