+++ src/Helpers.js
@@ -2353,93 +2353,6 @@ export function setTokenUsingIndexPrices(token, indexPrices, nativeTokenAddress)
   token.minPrice = indexPriceBn.mul(BASIS_POINTS_DIVISOR - halfSpreadBps).div(BASIS_POINTS_DIVISOR);
 }
 
-export function getInfoTokens(
-  tokens,
-  tokenBalances,
-  whitelistedTokens,
-  vaultTokenInfo,
-  fundingRateInfo,
-  vaultPropsLength,
-  indexPrices,
-  nativeTokenAddress
-) {
-  if (!vaultPropsLength) {
-    vaultPropsLength = 14;
-  }
-  const fundingRatePropsLength = 2;
-  const infoTokens = {};
-
-  for (let i = 0; i < tokens.length; i++) {
-    const token = JSON.parse(JSON.stringify(tokens[i]));
-    if (tokenBalances) {
-      token.balance = tokenBalances[i];
-    }
-    if (token.address === USDG_ADDRESS) {
-      token.minPrice = expandDecimals(1, USD_DECIMALS);
-      token.maxPrice = expandDecimals(1, USD_DECIMALS);
-    }
-    infoTokens[token.address] = token;
-  }
-
-  for (let i = 0; i < whitelistedTokens.length; i++) {
-    const token = JSON.parse(JSON.stringify(whitelistedTokens[i]));
-    if (vaultTokenInfo) {
-      token.poolAmount = vaultTokenInfo[i * vaultPropsLength];
-      token.reservedAmount = vaultTokenInfo[i * vaultPropsLength + 1];
-      token.availableAmount = token.poolAmount.sub(token.reservedAmount);
-      token.usdgAmount = vaultTokenInfo[i * vaultPropsLength + 2];
-      token.redemptionAmount = vaultTokenInfo[i * vaultPropsLength + 3];
-      token.weight = vaultTokenInfo[i * vaultPropsLength + 4];
-      token.bufferAmount = vaultTokenInfo[i * vaultPropsLength + 5];
-      token.maxUsdgAmount = vaultTokenInfo[i * vaultPropsLength + 6];
-      token.globalShortSize = vaultTokenInfo[i * vaultPropsLength + 7];
-      token.maxGlobalShortSize = vaultTokenInfo[i * vaultPropsLength + 8];
-      token.minPrice = vaultTokenInfo[i * vaultPropsLength + 9];
-      token.maxPrice = vaultTokenInfo[i * vaultPropsLength + 10];
-      token.guaranteedUsd = vaultTokenInfo[i * vaultPropsLength + 11];
-      token.maxPrimaryPrice = vaultTokenInfo[i * vaultPropsLength + 12];
-      token.minPrimaryPrice = vaultTokenInfo[i * vaultPropsLength + 13];
-
-      // save minPrice and maxPrice as setTokenUsingIndexPrices may override it
-      token.contractMinPrice = token.minPrice;
-      token.contractMaxPrice = token.maxPrice;
-
-      token.maxAvailableShort = bigNumberify(0);
-      if (token.maxGlobalShortSize.gt(0)) {
-        if (token.maxGlobalShortSize.gt(token.globalShortSize)) {
-          token.maxAvailableShort = token.maxGlobalShortSize.sub(token.globalShortSize);
-        }
-      }
-
-      if (token.maxUsdgAmount.eq(0)) {
-        token.maxUsdgAmount = DEFAULT_MAX_USDG_AMOUNT;
-      }
-
-      token.availableUsd = token.isStable
-        ? token.poolAmount.mul(token.minPrice).div(expandDecimals(1, token.decimals))
-        : token.availableAmount.mul(token.minPrice).div(expandDecimals(1, token.decimals));
-
-      token.managedUsd = token.availableUsd.add(token.guaranteedUsd);
-      token.managedAmount = token.managedUsd.mul(expandDecimals(1, token.decimals)).div(token.minPrice);
-
-      setTokenUsingIndexPrices(token, indexPrices, nativeTokenAddress);
-    }
-
-    if (fundingRateInfo) {
-      token.fundingRate = fundingRateInfo[i * fundingRatePropsLength];
-      token.cumulativeFundingRate = fundingRateInfo[i * fundingRatePropsLength + 1];
-    }
-
-    if (infoTokens[token.address]) {
-      token.balance = infoTokens[token.address].balance;
-    }
-
-    infoTokens[token.address] = token;
-  }
-
-  return infoTokens;
-}
-
 export const CHART_PERIODS = {
   "5m": 60 * 5,
   "15m": 60 * 15,
