+++ src/types/common.ts
@@ -1,8 +1,9 @@
 import { CHART_PERIODS } from "../Helpers";
+import { Web3ReactContextInterface } from "@web3-react/core/dist/types";
 
 export type ChainId = number; 
 export type TokenSymbol = string; 
 export type Period = keyof typeof CHART_PERIODS
 
 
-export type Library = any;
+export type Library = Web3ReactContextInterface["library"];
